Update TAL-Filter-2 and TAL-Reverb-2 to latest code

+++ ports/tal-filter-2-juce/src/Engine/Engine.h
@@ -115,8 +115,8 @@ public:
 
         if (this->filterType <= 7)
         {
-            this->filterHandlerL->setFiltertype(index);
-            this->filterHandlerR->setFiltertype(index);
+            this->filterHandlerL->setFiltertype((float)index);
+            this->filterHandlerR->setFiltertype((float)index);
         }
     }
